+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 0.2.0 / 2015-12-31
 ==================
-  * Added ES6 template string quotes \` support (\`templateUrl\`: \`/path/to/template.html\`)
+  * Added ES6 template string quotes \` support (templateUrl: \`/path/to/template.html\`)
 
 0.1.7 / 2015-12-31
 ==================